fix api and add bulk delete api

+++ src/Manufacturer/Design/design.controller.js
@@ -1,5 +1,7 @@
 import DesignService from "./design.service.js";
 import { getFullImagePath } from "./../../../utils/helper.js";
+import Prisma from './../../../db/prisma.js';
+
 import {
   UploadTo,
   deleteFile,
@@ -98,7 +100,7 @@ class DesignController {
         "sizeFrom",
         "sizeTo",
         "rate",
-        "colorValubodye",
+        "colors",
         "images",
       ]);
 
@@ -108,8 +110,12 @@ class DesignController {
         body.images = req.files.map((file) => file.filename);
         body.image = body.images[0];
       }
+      console.log(body);
 
-      body.colorValue = body?.colorValue?.split(",").map((v) => v.trim());
+      if (body.colors) {
+        body.colors = JSON.parse(body.colors);
+      }
+      console.log(body);
 
       const design = await DesignService.updateDesign(id, body);
 
@@ -121,9 +127,57 @@ class DesignController {
 
   static async deleteDesign(req, res) {
     try {
+      console.log("dynamic route hit with id:", req.params.id);
+      const design = await DesignService.getDesignById(req.params.id);
+
+      if (!design)
+        return res
+          .status(404)
+          .json({ status: false, message: "Design not found", data: null });
       await DesignService.deleteDesign(req.params.id);
+      if (design?.images[0]) {
+        deleteFile(design.images, "designs");
+      }
+
       res.success("Design deleted successfully");
     } catch (error) {
+      // console.log(error)
+      res.someThingWentWrong(error);
+    }
+  }
+
+  static async bulkdeleteDesign(req, res) {
+    try {
+      const { ids } = req.getBody(["ids"]);
+
+      const designs = await DesignService.getDesignByIds(ids);
+
+      if (designs.length === 0)
+        return res
+          .status(404)
+          .json({ status: false, message: "Design not found", data: null });
+
+      await Prisma.$transaction(async (tx) => {
+        await tx.design.deleteMany({
+          where: {
+            id: {
+              in: ids,
+            },
+          },
+        });
+      });
+
+      if (Array.isArray(designs)) {
+        designs.forEach((design) => {
+          if (Array.isArray(design.images)) {
+            deleteFile(design.images, "designs");
+          }
+        });
+      }
+
+      res.success("Design deleted successfully");
+    } catch (error) {
+      console.log(error)
       res.someThingWentWrong(error);
     }
   }
